+++ next.config.js
@@ -6,3 +6,14 @@ module.exports = {
     domains: ["wine21.speedgabia.com"],
   },
 };
+
+// module.exports = {
+//   webpack(config) {
+//     config.module.rules.push({
+//       test: /\.svg$/,
+//       use: ["@svgr/webpack"],
+//     });
+
+//     return config;
+//   },
+// };
