+++ package.json
@@ -29,6 +29,7 @@
   },
   "devDependencies": {
     "@serverless/cli": "^1.6.0",
+    "@svgr/webpack": "^6.4.0",
     "@types/node": "17.0.31",
     "@types/react": "18.0.8",
     "@types/react-dom": "18.0.3",
