+++ pages/_app.tsx
@@ -1,9 +1,9 @@
 import type { AppProps } from "next/app";
 import { ThemeProvider } from "styled-components";
-import GlobalStyles from "styles/GlobalStyles";
-import theme from "styles/theme";
-import Layout from "components/common/Layout/Layout";
-import SEO from "components/common/SEO/SEO";
+import GlobalStyles from "@/styles/GlobalStyles";
+import theme from "@/styles/theme";
+import Layout from "@/components/common/Layout/Layout";
+import SEO from "@/components/common/SEO/SEO";
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
