+++ src/constants/text-style.ts
@@ -1,4 +1,4 @@
-import theme from "styles/theme";
+import theme from "@/styles/theme";
 
 export type TextSize =
   | "xxs"
