+++ src/containers/searching/components/WineInfoDesc.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styled from "styled-components";
-import Text from "components/common/Text";
+import Text from "@/components/common/Text";
 
 const WineInfoDesc = ({ maker, country }) => {
   return (
