+++ src/containers/searching/components/WineInfoTags.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import styled from "styled-components";
-import { FLAG_BY_COUNTRY } from "types/Country";
-import { NAME_BY_WINE_TYPE } from "types/WineType";
-import Text from "components/common/Text";
+import { FLAG_BY_COUNTRY } from "@/types/Country";
+import { NAME_BY_WINE_TYPE } from "@/types/WineType";
+import Text from "@/components/common/Text";
 
 const WineInfoTags = ({ minAlcohol, maxAlcohol, country, type }) => {
   const isSameAlcohol = minAlcohol === maxAlcohol;
