+++ tsconfig.json
@@ -9,23 +9,13 @@
     "noEmit": true,
     "esModuleInterop": true,
     "module": "esnext",
-    "baseUrl": "./src",
     "moduleResolution": "node",
     "resolveJsonModule": true,
     "isolatedModules": true,
     "jsx": "preserve",
     "incremental": true,
-    "paths": {
-      "containers": ["src/containers/*"],
-      "components": ["src/components/*"],
-      "styles": ["src/styles/*"],
-      "hooks": ["src/hooks/*"],
-      "utils": ["src/utils/*"],
-      "constants": ["src/constants/*"],
-      "remotes": ["src/remotes/*"],
-      "models": ["src/models/*"],
-      "mocks": ["src/mocks/*"]
-    }
+    "baseUrl": ".",
+    "paths": { "@/*": ["src/*"] }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
